Added pre-calculated sin/cos lookup tables Added wrapRadians() function to keep radians from getting out of range. Was causing the player ship to rotate oddly.

+++ trunk/entity.cpp
@@ -183,7 +183,7 @@ void entity::run()
     mPos += mSpeed;
     mPos += mDrift;
     mAngle += mRotationRate;
-    mAngle = fmod((double)mAngle, 2.0f*PI);
+    mAngle = fmodf(mAngle, 2.0f*PI);
 
     // Keep it on the grid
     if (mGridBound)
